Move command compiler to core
